test: moved queries and mutations
