+++ __tests__/posts.ts
@@ -8030,637 +8030,3 @@ describe('query post awards', () => {
     ]);
   });
 });
-
-describe('mutation startPostBoost', () => {
-  const MUTATION = `
-    mutation StartPostBoost($postId: ID!, $duration: Int!, $budget: Int!) {
-      startPostBoost(postId: $postId, duration: $duration, budget: $budget) {
-        transactionId
-        balance {
-          amount
-        }
-      }
-    }
-  `;
-
-  const params = { postId: 'p1', duration: 7, budget: 5000 };
-
-  beforeEach(async () => {
-    isTeamMember = true; // TODO: remove when we are about to run production
-    await con.getRepository(Post).update({ id: 'p1' }, { authorId: '1' });
-  });
-
-  it('should not authorize when not logged in', () =>
-    testMutationErrorCode(
-      client,
-      { mutation: MUTATION, variables: params },
-      'UNAUTHENTICATED',
-    ));
-
-  it('should return an error if duration is less than 1', async () => {
-    loggedUser = '1';
-
-    return testMutationErrorCode(
-      client,
-      { mutation: MUTATION, variables: { ...params, duration: 0 } },
-      'GRAPHQL_VALIDATION_FAILED',
-    );
-  });
-
-  it('should return an error if duration is greater than 30', async () => {
-    loggedUser = '1';
-
-    return testMutationErrorCode(
-      client,
-      { mutation: MUTATION, variables: { ...params, duration: 31 } },
-      'GRAPHQL_VALIDATION_FAILED',
-    );
-  });
-
-  it('should return an error if budget is less than 1000', async () => {
-    loggedUser = '1';
-
-    return testMutationErrorCode(
-      client,
-      { mutation: MUTATION, variables: { ...params, budget: 999 } },
-      'GRAPHQL_VALIDATION_FAILED',
-    );
-  });
-
-  it('should return an error if budget is greater than 100000', async () => {
-    loggedUser = '1';
-
-    return testMutationErrorCode(
-      client,
-      { mutation: MUTATION, variables: { ...params, budget: 100001 } },
-      'GRAPHQL_VALIDATION_FAILED',
-    );
-  });
-
-  it('should return an error if budget is not divisible by 1000', async () => {
-    loggedUser = '1';
-
-    return testMutationErrorCode(
-      client,
-      { mutation: MUTATION, variables: { ...params, budget: 1500 } },
-      'GRAPHQL_VALIDATION_FAILED',
-    );
-  });
-
-  it('should return an error if post does not exist', async () => {
-    loggedUser = '1';
-
-    return testMutationErrorCode(
-      client,
-      { mutation: MUTATION, variables: { ...params, postId: 'nonexistent' } },
-      'NOT_FOUND',
-    );
-  });
-
-  it('should return an error if user is not the author or scout of the post', async () => {
-    loggedUser = '2'; // User 2 is not the author or scout of post p1
-
-    return testMutationErrorCode(
-      client,
-      { mutation: MUTATION, variables: params },
-      'NOT_FOUND',
-    );
-  });
-
-  it('should return an error if post is already boosted', async () => {
-    loggedUser = '1';
-    // Set the post as already boosted
-    await con.getRepository(Post).update(
-      { id: 'p1' },
-      {
-        flags: updateFlagsStatement<Post>({ campaignId: 'mock-id' }),
-      },
-    );
-
-    return testMutationErrorCode(
-      client,
-      { mutation: MUTATION, variables: params },
-      'GRAPHQL_VALIDATION_FAILED',
-    );
-  });
-
-  it('should handle skadi integration failure gracefully', async () => {
-    loggedUser = '1';
-
-    // Ensure the post is not boosted initially
-    await con.getRepository(Post).update(
-      { id: 'p1' },
-      {
-        flags: updateFlagsStatement<Post>({ campaignId: null }),
-      },
-    );
-
-    // Get initial transaction count
-    const initialTransactionCount = await con
-      .getRepository(UserTransaction)
-      .count({
-        where: { senderId: '1', referenceType: 'PostBoost' },
-      });
-
-    // This test validates that the mutation handles external service failures
-    // In a real scenario, this would test the error handling when skadi service is down
-    const res = await client.mutate(MUTATION, {
-      variables: { ...params, duration: 1, budget: 1000 },
-    });
-
-    // The mutation should fail due to external service issues
-    // but the validation logic should pass
-    expect(res.errors).toBeTruthy();
-
-    // Verify no new transactions were created
-    const finalTransactionCount = await con
-      .getRepository(UserTransaction)
-      .count({
-        where: { senderId: '1', referenceType: 'PostBoost' },
-      });
-    expect(finalTransactionCount).toBe(initialTransactionCount);
-
-    // Verify the boosted flag is still false
-    const post = await con.getRepository(Post).findOneBy({ id: 'p1' });
-    expect(post?.flags?.campaignId).toBeFalsy();
-  });
-
-  it('should handle transfer failure gracefully', async () => {
-    loggedUser = '1';
-
-    // Ensure the post is not boosted initially
-    await con.getRepository(Post).update(
-      { id: 'p1' },
-      {
-        flags: updateFlagsStatement<Post>({ campaignId: null }),
-      },
-    );
-
-    // Get initial transaction count
-    const initialTransactionCount = await con
-      .getRepository(UserTransaction)
-      .count({
-        where: { senderId: '1', referenceType: 'PostBoost' },
-      });
-
-    // This test validates that the mutation handles transfer failures
-    // In a real scenario, this would test the error handling when transfer fails
-    const res = await client.mutate(MUTATION, {
-      variables: { ...params, duration: 1, budget: 1000 },
-    });
-
-    // The mutation should fail due to external service issues
-    // but the validation logic should pass
-    expect(res.errors).toBeTruthy();
-
-    // Verify no new transactions were created
-    const finalTransactionCount = await con
-      .getRepository(UserTransaction)
-      .count({
-        where: { senderId: '1', referenceType: 'PostBoost' },
-      });
-    expect(finalTransactionCount).toBe(initialTransactionCount);
-
-    // Verify the boosted flag is still false
-    const post = await con.getRepository(Post).findOneBy({ id: 'p1' });
-    expect(post?.flags?.campaignId).toBeFalsy();
-  });
-
-  it('should verify no transactions are created when validation fails', async () => {
-    loggedUser = '1';
-
-    // Ensure the post is not boosted initially
-    await con.getRepository(Post).update(
-      { id: 'p1' },
-      {
-        flags: updateFlagsStatement<Post>({ campaignId: null }),
-      },
-    );
-
-    // Get initial transaction count
-    const initialTransactionCount = await con
-      .getRepository(UserTransaction)
-      .count({
-        where: { senderId: '1', referenceType: 'PostBoost' },
-      });
-
-    // Try to boost with invalid parameters
-    const res = await client.mutate(MUTATION, {
-      variables: { ...params, duration: 0 }, // Invalid duration
-    });
-
-    // Should fail validation
-    expect(res.errors).toBeTruthy();
-
-    // Verify no new transactions were created
-    const finalTransactionCount = await con
-      .getRepository(UserTransaction)
-      .count({
-        where: { senderId: '1', referenceType: 'PostBoost' },
-      });
-    expect(finalTransactionCount).toBe(initialTransactionCount);
-
-    // Verify the boosted flag is still false
-    const post = await con.getRepository(Post).findOneBy({ id: 'p1' });
-    expect(post?.flags?.campaignId).toBeFalsy();
-  });
-
-  it('should verify no transactions are created when post does not exist', async () => {
-    loggedUser = '1';
-
-    // Get initial transaction count
-    const initialTransactionCount = await con
-      .getRepository(UserTransaction)
-      .count({
-        where: { senderId: '1', referenceType: 'PostBoost' },
-      });
-
-    // Try to boost a non-existent post
-    const res = await client.mutate(MUTATION, {
-      variables: { ...params, postId: 'nonexistent' },
-    });
-
-    // Should fail with not found error
-    expect(res.errors).toBeTruthy();
-
-    // Verify no new transactions were created
-    const finalTransactionCount = await con
-      .getRepository(UserTransaction)
-      .count({
-        where: { senderId: '1', referenceType: 'PostBoost' },
-      });
-    expect(finalTransactionCount).toBe(initialTransactionCount);
-
-    // Verify the original post's boosted flag is unchanged
-    const post = await con.getRepository(Post).findOneBy({ id: 'p1' });
-    expect(post?.flags?.campaignId).toBeFalsy();
-  });
-
-  it('should verify no transactions are created when user is not authorized', async () => {
-    loggedUser = '2'; // User 2 is not the author or scout of post p1
-
-    // Ensure the post is not boosted initially
-    await con.getRepository(Post).update(
-      { id: 'p1' },
-      {
-        flags: updateFlagsStatement<Post>({ campaignId: null }),
-      },
-    );
-
-    // Get initial transaction count for user 2
-    const initialTransactionCount = await con
-      .getRepository(UserTransaction)
-      .count({
-        where: { senderId: '2', referenceType: 'PostBoost' },
-      });
-
-    // Try to boost without authorization
-    const res = await client.mutate(MUTATION, {
-      variables: params,
-    });
-
-    // Should fail with not found error
-    expect(res.errors).toBeTruthy();
-
-    // Verify no new transactions were created for user 2
-    const finalTransactionCount = await con
-      .getRepository(UserTransaction)
-      .count({
-        where: { senderId: '2', referenceType: 'PostBoost' },
-      });
-    expect(finalTransactionCount).toBe(initialTransactionCount);
-
-    // Verify the boosted flag is still false
-    const post = await con.getRepository(Post).findOneBy({ id: 'p1' });
-    expect(post?.flags?.campaignId).toBeFalsy();
-  });
-
-  it('should verify no transactions are created when post is already boosted', async () => {
-    loggedUser = '1';
-
-    // Set the post as already boosted
-    await con.getRepository(Post).update(
-      { id: 'p1' },
-      {
-        flags: updateFlagsStatement<Post>({ campaignId: 'mock-id' }),
-      },
-    );
-
-    // Get initial transaction count
-    const initialTransactionCount = await con
-      .getRepository(UserTransaction)
-      .count({
-        where: { senderId: '1', referenceType: 'PostBoost' },
-      });
-
-    // Try to boost an already boosted post
-    const res = await client.mutate(MUTATION, {
-      variables: params,
-    });
-
-    // Should fail with validation error
-    expect(res.errors).toBeTruthy();
-
-    // Verify no new transactions were created
-    const finalTransactionCount = await con
-      .getRepository(UserTransaction)
-      .count({
-        where: { senderId: '1', referenceType: 'PostBoost' },
-      });
-    expect(finalTransactionCount).toBe(initialTransactionCount);
-
-    // Verify the boosted flag remains true (unchanged)
-    const post = await con.getRepository(Post).findOneBy({ id: 'p1' });
-    expect(post?.flags?.campaignId).toBe('mock-id');
-  });
-});
-
-describe('mutation cancelPostBoost', () => {
-  const MUTATION = `
-    mutation CancelPostBoost($postId: ID!) {
-      cancelPostBoost(postId: $postId) {
-        _
-      }
-    }
-  `;
-
-  const params = { postId: 'p1' };
-
-  beforeEach(async () => {
-    isTeamMember = true; // TODO: remove when we are about to run production
-    await con.getRepository(Post).update(
-      { id: 'p1' },
-      {
-        authorId: '1',
-        flags: updateFlagsStatement<Post>({ campaignId: 'mock-id' }),
-      },
-    );
-  });
-
-  it('should not authorize when not logged in', () =>
-    testMutationErrorCode(
-      client,
-      { mutation: MUTATION, variables: params },
-      'UNAUTHENTICATED',
-    ));
-
-  it('should return an error if post does not exist', async () => {
-    loggedUser = '1';
-
-    return testMutationErrorCode(
-      client,
-      { mutation: MUTATION, variables: { postId: 'nonexistent' } },
-      'NOT_FOUND',
-    );
-  });
-
-  it('should return an error if user is not the author or scout of the post', async () => {
-    loggedUser = '2'; // User 2 is not the author or scout of post p1
-
-    return testMutationErrorCode(
-      client,
-      { mutation: MUTATION, variables: params },
-      'NOT_FOUND',
-    );
-  });
-
-  it('should return an error if post is not currently boosted', async () => {
-    loggedUser = '1';
-    // Ensure the post is not boosted
-    await con.getRepository(Post).update(
-      { id: 'p1' },
-      {
-        flags: updateFlagsStatement<Post>({ campaignId: null }),
-      },
-    );
-
-    return testMutationErrorCode(
-      client,
-      { mutation: MUTATION, variables: params },
-      'GRAPHQL_VALIDATION_FAILED',
-    );
-  });
-
-  it('should successfully cancel post boost when post is boosted', async () => {
-    loggedUser = '1';
-
-    const res = await client.mutate(MUTATION, {
-      variables: params,
-    });
-
-    expect(res.errors).toBeFalsy();
-    expect(res.data.cancelPostBoost._).toBe(true);
-
-    // Verify the boosted flag is now false
-    const post = await con.getRepository(Post).findOneBy({ id: 'p1' });
-    expect(post?.flags?.campaignId).toBeFalsy();
-  });
-
-  it('should verify boosted flag remains unchanged when validation fails', async () => {
-    loggedUser = '1';
-
-    // Ensure the post is not boosted initially
-    await con.getRepository(Post).update(
-      { id: 'p1' },
-      {
-        flags: updateFlagsStatement<Post>({ campaignId: null }),
-      },
-    );
-
-    // Try to cancel boost on a non-boosted post
-    const res = await client.mutate(MUTATION, {
-      variables: params,
-    });
-
-    // Should fail validation
-    expect(res.errors).toBeTruthy();
-
-    // Verify the boosted flag remains false (unchanged)
-    const post = await con.getRepository(Post).findOneBy({ id: 'p1' });
-    expect(post?.flags?.campaignId).toBeFalsy();
-  });
-
-  it('should verify boosted flag remains unchanged when post does not exist', async () => {
-    loggedUser = '1';
-
-    // Try to cancel boost on a non-existent post
-    const res = await client.mutate(MUTATION, {
-      variables: { postId: 'nonexistent' },
-    });
-
-    // Should fail with not found error
-    expect(res.errors).toBeTruthy();
-
-    // Verify the original post's boosted flag is unchanged
-    const post = await con.getRepository(Post).findOneBy({ id: 'p1' });
-    // The boosted flag should be whatever it was before (false by default)
-    expect(post?.flags?.campaignId).toBe('mock-id');
-  });
-
-  it('should verify boosted flag remains unchanged when user is not authorized', async () => {
-    loggedUser = '2'; // User 2 is not the author or scout of post p1
-
-    // Try to cancel boost without authorization
-    const res = await client.mutate(MUTATION, {
-      variables: params,
-    });
-
-    // Should fail with not found error
-    expect(res.errors).toBeTruthy();
-
-    // Verify the boosted flag remains true (unchanged)
-    const post = await con.getRepository(Post).findOneBy({ id: 'p1' });
-    expect(post?.flags?.campaignId).toBe('mock-id');
-  });
-
-  it('should work for post scout as well as author', async () => {
-    loggedUser = '1';
-
-    // Set user as scout instead of author
-    await con.getRepository(Post).update(
-      { id: 'p1' },
-      {
-        authorId: '2', // Different author
-        scoutId: '1', // Current user is scout
-      },
-    );
-
-    const res = await client.mutate(MUTATION, {
-      variables: params,
-    });
-
-    expect(res.errors).toBeFalsy();
-    expect(res.data.cancelPostBoost._).toBe(true);
-
-    // Verify the boosted flag is now false
-    const post = await con.getRepository(Post).findOneBy({ id: 'p1' });
-    expect(post?.flags?.campaignId).toBeFalsy();
-  });
-});
-
-describe('query boostEstimatedReach', () => {
-  const QUERY = `
-    query BoostEstimatedReach($postId: ID!, $duration: Int!, $budget: Int!) {
-      boostEstimatedReach(postId: $postId, duration: $duration, budget: $budget) {
-        estimatedReach {
-          min
-          max
-        }
-      }
-    }
-  `;
-
-  const params = { postId: 'p1', duration: 7, budget: 5000 };
-
-  beforeEach(async () => {
-    isTeamMember = true; // TODO: remove when we are about to run production
-    await con.getRepository(Post).update({ id: 'p1' }, { authorId: '1' });
-  });
-
-  it('should not authorize when not logged in', () =>
-    testQueryErrorCode(
-      client,
-      { query: QUERY, variables: params },
-      'UNAUTHENTICATED',
-    ));
-
-  it('should return an error if duration is less than 1', async () => {
-    loggedUser = '1';
-
-    return testQueryErrorCode(
-      client,
-      { query: QUERY, variables: { ...params, duration: 0 } },
-      'GRAPHQL_VALIDATION_FAILED',
-    );
-  });
-
-  it('should return an error if duration is greater than 30', async () => {
-    loggedUser = '1';
-
-    return testQueryErrorCode(
-      client,
-      { query: QUERY, variables: { ...params, duration: 31 } },
-      'GRAPHQL_VALIDATION_FAILED',
-    );
-  });
-
-  it('should return an error if budget is less than 1000', async () => {
-    loggedUser = '1';
-
-    return testQueryErrorCode(
-      client,
-      { query: QUERY, variables: { ...params, budget: 999 } },
-      'GRAPHQL_VALIDATION_FAILED',
-    );
-  });
-
-  it('should return an error if budget is greater than 100000', async () => {
-    loggedUser = '1';
-
-    return testQueryErrorCode(
-      client,
-      { query: QUERY, variables: { ...params, budget: 100001 } },
-      'GRAPHQL_VALIDATION_FAILED',
-    );
-  });
-
-  it('should return an error if budget is not divisible by 1000', async () => {
-    loggedUser = '1';
-
-    return testQueryErrorCode(
-      client,
-      { query: QUERY, variables: { ...params, budget: 1500 } },
-      'GRAPHQL_VALIDATION_FAILED',
-    );
-  });
-
-  it('should return an error if post does not exist', async () => {
-    loggedUser = '1';
-
-    return testQueryErrorCode(
-      client,
-      { query: QUERY, variables: { ...params, postId: 'nonexistent' } },
-      'NOT_FOUND',
-    );
-  });
-
-  it('should return an error if user is not the author or scout of the post', async () => {
-    loggedUser = '2'; // User 2 is not the author or scout of post p1
-
-    return testQueryErrorCode(
-      client,
-      { query: QUERY, variables: params },
-      'NOT_FOUND',
-    );
-  });
-
-  it('should return an error if post is already boosted', async () => {
-    loggedUser = '1';
-    // Set the post as already boosted
-    await con.getRepository(Post).update(
-      { id: 'p1' },
-      {
-        flags: updateFlagsStatement<Post>({ campaignId: 'mock-id' }),
-      },
-    );
-
-    return testQueryErrorCode(
-      client,
-      { query: QUERY, variables: params },
-      'GRAPHQL_VALIDATION_FAILED',
-    );
-  });
-
-  it('should the response returned by skadi client', async () => {
-    loggedUser = '1';
-
-    const res = await client.query(QUERY, {
-      variables: { ...params, duration: 1, budget: 1000 },
-    });
-
-    expect(res.data.boostEstimatedReach.estimatedReach).toEqual({
-      max: 120,
-      min: 80,
-    });
-  });
-});
